Add perk system bolt-on pack (PlayerPerkTracker, PerkHUD, README)
Adds the initial perk bolt-on pack to the project.  
Includes two new scripts (PlayerPerkTracker and PerkHUD) and a detailed
README explaining how to integrate these features directly into the Unity
project. This branch is intended to be a safe reference layer that the
rest of the team can copy into Assets/Scripts and patch existing files
as outlined in the documentation.